backend/requirements.txt
Flask
Flask-Cors
gunicorn
pandas
psycopg2-binary
requests
scikit-learn
Werkzeug